requirements.txt
fastapi
uvicorn
openai-whisper
torch
torchaudio
torchvision
ffmpeg-python
